Backend/requirements.txt
fastapi
pydantic
langchain_google_genai
langchain_core
uvicorn
python-dotenv
